ci: retest smoketest workflow
