+++ package.json
@@ -9,6 +9,7 @@
     "lint": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
     "prepare": "husky",
+    "smoketest": "tsx .github/smoketest/selector-check.ts",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
@@ -56,6 +57,7 @@
     "playwright": "^1.52.0",
     "prettier": "^3.5.3",
     "semantic-release": "^24.2.3",
+    "tsx": "^4.19.4",
     "typescript": "^5.8.3",
     "vite": "^6.3.2",
     "vite-plugin-static-copy": "^2.3.1",
